refactor: KDE: Add KwinScripts struct
