--- src/client/kde/kwin-script-one-off.js
@@ -1,22 +1,0 @@
-function notifyActiveWindow(client) {
-    if (!client) {
-        // Ignore when there is no active window.
-        return;
-    }
-
-    callDBus(
-        "com.k0kubun.Xremap",
-        "/com/k0kubun/Xremap",
-        "com.k0kubun.Xremap",
-        "NotifyActiveWindow",
-        "caption" in client ? client.caption : "",
-        "resourceClass" in client ? client.resourceClass : "",
-    );
-}
-
-// Send info about current window right away
-let active_window = workspace.activeClient
-    ? workspace.activeClient
-    : workspace.activeWindow;
-
-notifyActiveWindow(active_window);
