+++ src/client/kde/mod.rs
@@ -1,5 +1,6 @@
 mod adhoc_script_handler;
 mod kde_client;
+mod kwin_scripts;
 mod plugin_script_handler;
 
 pub use kde_client::KdeClient;
